+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 * Added flow to verify account address.
+* Added support for register data transaction.
 
 ## 2.0.3 - 2021-11-29